SubscriberEntityToConfigurationMapper moved to a Application.Infrastructure project
