+++ src/CaptainHook.Application.Infrastructure/CaptainHook.Application.Infrastructure.csproj
@@ -5,7 +5,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.ServiceFabric.Services" Version="4.1.428" />
+    <PackageReference Include="Microsoft.ServiceFabric.Services" Version="4.0.470" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CaptainHook.Common\CaptainHook.Common.csproj" />
+    <ProjectReference Include="..\CaptainHook.Domain\CaptainHook.Domain.csproj" />
   </ItemGroup>
 
 </Project>
