+++ src/CaptainHook.DirectorService/CaptainHook.DirectorService.csproj
@@ -28,6 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\CaptainHook.Application.Infrastructure\CaptainHook.Application.Infrastructure.csproj" />
     <ProjectReference Include="..\CaptainHook.Application\CaptainHook.Application.csproj" />
     <ProjectReference Include="..\CaptainHook.Common\CaptainHook.Common.csproj" />
     <ProjectReference Include="..\CaptainHook.Domain\CaptainHook.Domain.csproj" />
